Changed: Made Multiple Changes
Changed the name of the Plugin from MetaCommandsBlocker to PluginCommandsBlocker as it was more descriptive of the plugin's current functionality. So remember to remove the MetaCommandsBlocker if you had the previous version when updating to this one.

Changed the folder structure and general directory layout for easier installation purposes.

Updated the ReadMe file to be more concise and for easier readability.

Added a minimum API version check, to ensure that the server is running at least version 96 of CounterStrikeSharp when using the plugin.

Implemented a check to verify that the user who tries to access the plugin information has root access, to prevent administrators with the generic flag from accessing the server's plugin information.